Move test functions into meta module
