+++ saleor/graphql/meta/tests/test_meta_mutations.py
@@ -8,27 +8,7 @@ import pytest
 from ....core.error_codes import MetadataErrorCode
 from ....core.models import ModelWithMetadata
 from ....invoice.models import Invoice
-from ...tests.utils import assert_no_permission
-from .utils import (
-    DELETE_PRIVATE_METADATA_MUTATION,
-    UPDATE_PRIVATE_METADATA_MUTATION,
-    execute_clear_private_metadata_for_item,
-    execute_clear_private_metadata_for_multiple_items,
-    execute_clear_public_metadata_for_item,
-    execute_clear_public_metadata_for_multiple_items,
-    execute_update_private_metadata_for_item,
-    execute_update_private_metadata_for_multiple_items,
-    execute_update_public_metadata_for_item,
-    execute_update_public_metadata_for_multiple_items,
-    item_contains_multiple_proper_private_metadata,
-    item_contains_multiple_proper_public_metadata,
-    item_contains_proper_private_metadata,
-    item_contains_proper_public_metadata,
-    item_without_multiple_private_metadata,
-    item_without_multiple_public_metadata,
-    item_without_private_metadata,
-    item_without_public_metadata,
-)
+from ...tests.utils import assert_no_permission, get_graphql_content
 
 PRIVATE_KEY = "private_key"
 PRIVATE_VALUE = "private_vale"
@@ -37,6 +17,110 @@ PUBLIC_KEY = "key"
 PUBLIC_KEY2 = "key2"
 PUBLIC_VALUE = "value"
 PUBLIC_VALUE2 = "value2"
+
+
+UPDATE_PUBLIC_METADATA_MUTATION = """
+mutation UpdatePublicMetadata($id: ID!, $input: [MetadataInput!]!) {
+    updateMetadata(
+        id: $id
+        input: $input
+    ) {
+        metadataErrors{
+            field
+            code
+            message
+        }
+        item {
+            metadata{
+                key
+                value
+            }
+            ...on %s{
+                id
+            }
+        }
+    }
+}
+"""
+
+
+def execute_update_public_metadata_for_item(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+):
+    variables = {
+        "id": item_id,
+        "input": [{"key": key, "value": value}],
+    }
+
+    response = client.post_graphql(
+        UPDATE_PUBLIC_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def execute_update_public_metadata_for_multiple_items(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    variables = {
+        "id": item_id,
+        "input": [{"key": key, "value": value}, {"key": key2, "value": value2}],
+    }
+
+    response = client.post_graphql(
+        UPDATE_PUBLIC_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def item_contains_proper_public_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return item.get_value_from_metadata(key) == value
+
+
+def item_contains_multiple_proper_public_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return all(
+        [
+            item.get_value_from_metadata(key) == value,
+            item.get_value_from_metadata(key2) == value2,
+        ]
+    )
 
 
 def test_add_public_metadata_for_customer_as_staff(
@@ -617,6 +701,101 @@ def test_update_public_metadata_for_item_without_meta(api_client, address):
     errors = response["data"]["updateMetadata"]["metadataErrors"]
     assert errors[0]["field"] == "id"
     assert errors[0]["code"] == MetadataErrorCode.NOT_FOUND.name
+
+
+DELETE_PUBLIC_METADATA_MUTATION = """
+mutation DeletePublicMetadata($id: ID!, $keys: [String!]!) {
+    deleteMetadata(
+        id: $id
+        keys: $keys
+    ) {
+        metadataErrors{
+            field
+            code
+        }
+        item {
+            metadata{
+                key
+                value
+            }
+            ...on %s{
+                id
+            }
+        }
+    }
+}
+"""
+
+
+def execute_clear_public_metadata_for_item(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PUBLIC_KEY,
+):
+    variables = {
+        "id": item_id,
+        "keys": [key],
+    }
+
+    response = client.post_graphql(
+        DELETE_PUBLIC_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def execute_clear_public_metadata_for_multiple_items(
+    client, permissions, item_id, item_type, key=PUBLIC_KEY, key2=PUBLIC_KEY2
+):
+    variables = {
+        "id": item_id,
+        "keys": [key, key2],
+    }
+
+    response = client.post_graphql(
+        DELETE_PUBLIC_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def item_without_public_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return item.get_value_from_metadata(key) != value
+
+
+def item_without_multiple_public_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return all(
+        [
+            item.get_value_from_metadata(key) != value,
+            item.get_value_from_metadata(key2) != value2,
+        ]
+    )
 
 
 def test_delete_public_metadata_for_customer_as_staff(
@@ -1212,6 +1391,109 @@ def test_delete_public_metadata_for_one_key(api_client, checkout):
     )
 
 
+UPDATE_PRIVATE_METADATA_MUTATION = """
+mutation UpdatePrivateMetadata($id: ID!, $input: [MetadataInput!]!) {
+    updatePrivateMetadata(
+        id: $id
+        input: $input
+    ) {
+        metadataErrors{
+            field
+            code
+        }
+        item {
+            privateMetadata{
+                key
+                value
+            }
+            ...on %s{
+                id
+            }
+        }
+    }
+}
+"""
+
+
+def execute_update_private_metadata_for_item(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PRIVATE_KEY,
+    value=PRIVATE_VALUE,
+):
+    variables = {
+        "id": item_id,
+        "input": [{"key": key, "value": value}],
+    }
+
+    response = client.post_graphql(
+        UPDATE_PRIVATE_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def execute_update_private_metadata_for_multiple_items(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    variables = {
+        "id": item_id,
+        "input": [{"key": key, "value": value}, {"key": key2, "value": value2}],
+    }
+
+    response = client.post_graphql(
+        UPDATE_PRIVATE_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def item_contains_proper_private_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PRIVATE_KEY,
+    value=PRIVATE_VALUE,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return item.get_value_from_private_metadata(key) == value
+
+
+def item_contains_multiple_proper_private_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return all(
+        [
+            item.get_value_from_private_metadata(key) == value,
+            item.get_value_from_private_metadata(key2) == value2,
+        ]
+    )
+
+
 def test_add_private_metadata_for_customer_as_staff(
     staff_api_client, permission_manage_users, customer_user
 ):
@@ -1786,6 +2068,101 @@ def test_update_private_metadata_for_item_without_meta(api_client, address):
     errors = response["data"]["updatePrivateMetadata"]["metadataErrors"]
     assert errors[0]["field"] == "id"
     assert errors[0]["code"] == MetadataErrorCode.NOT_FOUND.name
+
+
+DELETE_PRIVATE_METADATA_MUTATION = """
+mutation DeletePrivateMetadata($id: ID!, $keys: [String!]!) {
+    deletePrivateMetadata(
+        id: $id
+        keys: $keys
+    ) {
+        metadataErrors{
+            field
+            code
+        }
+        item {
+            privateMetadata{
+                key
+                value
+            }
+            ...on %s{
+                id
+            }
+        }
+    }
+}
+"""
+
+
+def execute_clear_private_metadata_for_item(
+    client,
+    permissions,
+    item_id,
+    item_type,
+    key=PRIVATE_KEY,
+):
+    variables = {
+        "id": item_id,
+        "keys": [key],
+    }
+
+    response = client.post_graphql(
+        DELETE_PRIVATE_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def execute_clear_private_metadata_for_multiple_items(
+    client, permissions, item_id, item_type, key=PUBLIC_KEY, key2=PUBLIC_KEY2
+):
+    variables = {
+        "id": item_id,
+        "keys": [key, key2],
+    }
+
+    response = client.post_graphql(
+        DELETE_PUBLIC_METADATA_MUTATION % item_type,
+        variables,
+        permissions=[permissions] if permissions else None,
+    )
+    response = get_graphql_content(response)
+    return response
+
+
+def item_without_private_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PRIVATE_KEY,
+    value=PRIVATE_VALUE,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return item.get_value_from_private_metadata(key) != value
+
+
+def item_without_multiple_private_metadata(
+    item_from_response,
+    item,
+    item_id,
+    key=PUBLIC_KEY,
+    value=PUBLIC_VALUE,
+    key2=PUBLIC_KEY2,
+    value2=PUBLIC_VALUE2,
+):
+    if item_from_response["id"] != item_id:
+        return False
+    item.refresh_from_db()
+    return all(
+        [
+            item.get_value_from_private_metadata(key) != value,
+            item.get_value_from_private_metadata(key2) != value2,
+        ]
+    )
 
 
 def test_delete_private_metadata_for_customer_as_staff(
@@ -2412,4 +2789,76 @@ def test_delete_private_metadata_for_one_key(
         checkout,
         checkout_id,
         key="to_clear",
+    )
+
+
+def test_add_public_metadata_for_warehouse(
+    staff_api_client, permission_manage_products, warehouse
+):
+    # given
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    # when
+    response = execute_update_public_metadata_for_item(
+        staff_api_client, permission_manage_products, warehouse_id, "Warehouse"
+    )
+
+    # then
+    assert item_contains_proper_public_metadata(
+        response["data"]["updateMetadata"]["item"], warehouse, warehouse_id
+    )
+
+
+def test_delete_public_metadata_for_warehouse(
+    staff_api_client, permission_manage_products, warehouse
+):
+    # given
+    warehouse.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    warehouse.save(update_fields=["metadata"])
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    # when
+    response = execute_clear_public_metadata_for_item(
+        staff_api_client, permission_manage_products, warehouse_id, "Warehouse"
+    )
+
+    # then
+    assert item_without_public_metadata(
+        response["data"]["deleteMetadata"]["item"], warehouse, warehouse_id
+    )
+
+
+def test_add_private_metadata_for_warehouse(
+    staff_api_client, permission_manage_products, warehouse
+):
+    # given
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    # when
+    response = execute_update_private_metadata_for_item(
+        staff_api_client, permission_manage_products, warehouse_id, "Warehouse"
+    )
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"], warehouse, warehouse_id
+    )
+
+
+def test_delete_private_metadata_for_warehouse(
+    staff_api_client, permission_manage_products, warehouse
+):
+    # given
+    warehouse.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
+    warehouse.save(update_fields=["private_metadata"])
+    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
+
+    # when
+    response = execute_clear_private_metadata_for_item(
+        staff_api_client, permission_manage_products, warehouse_id, "Warehouse"
+    )
+
+    # then
+    assert item_without_private_metadata(
+        response["data"]["deletePrivateMetadata"]["item"], warehouse, warehouse_id
     )
